// APITypes/ProjectInfo.cs.d.ts
export declare interface projectInfo {
	title: string;
	description: string;
	projectId: string;
	minutes: number;
	highlights: number;
	events: number;
	authorName: string;
	authorDescription: string;
	authorId: string;
	playbackDurationMs: number;
	avatarUri: string;
	socialAccount1: string;
	socialAccount2: string;
	socialAccount3: string;
	merchandiseLink1: string;
	merchandiseLink2: string;
	merchandiseVideoUrl: string;
	logoUrl: string;
	enableMerchandiseScreen: boolean;
}